Develop

+++ frontend/RollingEats/src/pages/AdminUserScreen.jsx
@@ -7,6 +7,7 @@ import EditUserModal from "../components/EditUserModal";
 
 const AdminUserScreen = () => {
 	const [users, setUsers] = useState([]);
+  const [loading, setLoading] = useState(true);
 
   const MySwal = withReactContent(Swal);
 
@@ -28,11 +29,15 @@ const AdminUserScreen = () => {
   const changeStatus = async (id) => {
     try {
       const response = await getUserById(id);
-      console.log(response)
       const user = response.user;
-      user.status = !user.status;
-      const result = await editUserById(id, user);
-      fetchData();
+      if(user.role === "ADMIN_ROLE"){
+        MySwal.fire("No se puede bloquear a un administrador", "", "info");
+        return;
+      }else{
+        user.status = !user.status;
+        const result = await editUserById(id, user);
+        fetchData();
+      }
     } catch (e) {
       console.error(e);
     }
@@ -40,23 +45,29 @@ const AdminUserScreen = () => {
 
   // Block Order
   const blockUser = async (id) => {
-    MySwal.fire({
-      title: `¿Está seguro de que quiere inactivar el usuario con ID ${id}?`,
-      showDenyButton: true,
-      showCancelButton: false,
-      confirmButtonText: "Sí",
-      denyButtonText: `No`,
-    }).then((result) => {
-      if (result.isConfirmed) {
-        deleteUserById(id).then((result) => {
-          fetchData();
-          console.log(result)
-          MySwal.fire("", `${result.message}`, "success");
-        });
-      } else if (result.isDenied) {
-        MySwal.fire("El usuario no pudo ser inactivado", "", "info");
-      }
-    });
+    const response = await getUserById(id);
+    if(response.user.role === "ADMIN_ROLE"){
+      MySwal.fire("No se puede bloquear a un administrador", "", "info");
+      return;
+    }else{
+      MySwal.fire({
+        title: `¿Está seguro de que quiere inactivar el usuario con ID ${id}?`,
+        showDenyButton: true,
+        showCancelButton: false,
+        confirmButtonText: "Sí",
+        denyButtonText: `No`,
+      }).then((result) => {
+        if (result.isConfirmed) {
+          deleteUserById(id).then((result) => {
+            fetchData();
+            console.log(result)
+            MySwal.fire("", `${result.message}`, "success");
+          });
+        } else if (result.isDenied) {
+          MySwal.fire("El usuario no pudo ser inactivado", "", "info");
+        }
+      });
+    }
   };
 
 
@@ -68,6 +79,7 @@ const AdminUserScreen = () => {
     try {
       const response = await getUsersWithoutStatus();
       setUsers(response.users);
+      setLoading(false);
     } catch (e) {
       console.error(error);
     }
@@ -76,6 +88,14 @@ const AdminUserScreen = () => {
 	return (
     <>
     <br /><br /><br />
+    {loading == true ? (
+				<>
+				<div class="spinner-border custom-spinner" role="status">
+					<span class="visually-hidden">Loading...</span>
+				</div>
+				<br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br /><br />
+				</>
+			) : (
 		<div className="m-5 table-responsive">
 			<table className="table table-hover table-striped table-bordered">
 				<thead className="bg-thead">
@@ -109,6 +129,7 @@ const AdminUserScreen = () => {
 				</tbody>
 			</table>
 		</div>
+      )}
     {show && (
         <EditUserModal show={show} handleClose={handleClose} uid={uid} />
       )}
